delete specedic tasks and delette all

+++ delete.php
@@ -1,26 +1,27 @@
 <?php
-
 if(isset($_GET["id"]) ){
-
-
     $id = $_GET["id"];
+    $servername = "localhost";
+    $username = "root";
+    $password = "";
+    $database = "todolist";
 
+    $connection = new mysqli($servername, $username, $password, $database);
 
-    
-$servername = "localhost";
-$username = "root";
-$password = "";
-$database = "todolist";
+    if ($connection->connect_error) {
+        die("Connection failed: " . $connection->connect_error);
+    }
 
-$connection = new mysqli($servername, $username, $password, $database);
+    $sql = "DELETE FROM task WHERE id=$id";
 
-
-$sql = "DELETE FROM task WHERE id=$id";
-$connection->query($sql);
-
+    if ($connection->query($sql) === TRUE) {
+        header("Location: ./index.php/");
+        exit();
+    } else {
+        echo "Error deleting record: " . $connection->error;
+    }
+    $connection->close();
+} else {
+    echo "No task ID specified.";
 }
-header("location:./index.php");
-exit;
-
-
 ?>
